luocaca_company  ：         删除     无用资源。。二次commit   保存节点

+++ app/src/main/java/com/autoscrollview/widget/AutoScrollViewPager.java
@@ -100,7 +100,7 @@ public class AutoScrollViewPager extends InsideViewPager {
      */
     public void startAutoScroll() {
         isAutoScroll = true;
-        sendScrollMessage((long) (interval));
+        sendScrollMessage(interval);
         // sendScrollMessage((long) (interval + scroller.getDuration()/
         // autoScrollFactor * swipeScrollFactor));
     }
