+++ app/src/main/java/com/hldj/hmyg/StoreActivity.java
@@ -222,7 +222,7 @@ public class StoreActivity extends NeedSwipeBackActivity implements
                 LinearLayout.LayoutParams.MATCH_PARENT,
                 LinearLayout.LayoutParams.WRAP_CONTENT);
         WindowManager wm = this.getWindowManager();
-        l_params.height = (int) (wm.getDefaultDisplay().getWidth() * 1 / 2);
+        l_params.height = wm.getDefaultDisplay().getWidth() * 1 / 2;
         iv_banner.setLayoutParams(l_params);
         tv_store_name = (TextView) findViewById(R.id.tv_store_name);
         tv_plant_type = (TextView) findViewById(R.id.tv_plant_type);
