+++ app/src/main/java/com/hldj/hmyg/adapter/ThematicAdapter.java
@@ -40,7 +40,7 @@ public class ThematicAdapter extends BaseAdapter {
         l_params = new LinearLayout.LayoutParams(LayoutParams.MATCH_PARENT,
                 LayoutParams.WRAP_CONTENT);
         WindowManager wm = ((Activity) context).getWindowManager();
-        l_params.height = (int) (wm.getDefaultDisplay().getWidth() * 9 / 20);
+        l_params.height = wm.getDefaultDisplay().getWidth() * 9 / 20;
     }
 
     @Override
