+++ app/src/main/java/com/hldj/hmyg/buy/adapter/SaveForCartSeedlingAdapter.java
@@ -46,7 +46,7 @@ public class SaveForCartSeedlingAdapter extends BaseAdapter {
 		l_params = new LinearLayout.LayoutParams(LayoutParams.MATCH_PARENT,
 				LayoutParams.WRAP_CONTENT);
 		WindowManager wm = ((Activity) context).getWindowManager();
-		l_params.height = (int) (wm.getDefaultDisplay().getWidth() * 1 / 3);
+		l_params.height = wm.getDefaultDisplay().getWidth() * 1 / 3;
 	}
 
 	@Override
