+++ app/src/main/java/com/hldj/hmyg/buyer/Ui/OnlyDirstreetWheelDialogF.java
@@ -70,7 +70,7 @@ public class OnlyDirstreetWheelDialogF extends DialogFragment implements OnWheel
         super.onStart();
         WindowManager.LayoutParams lp = getDialog().getWindow().getAttributes();
         lp.gravity = Gravity.BOTTOM;
-        lp.width = (int) (getDialog().getWindow().getWindowManager().getDefaultDisplay().getWidth());
+        lp.width = getDialog().getWindow().getWindowManager().getDefaultDisplay().getWidth();
         getDialog().getWindow().setGravity(Gravity.BOTTOM);
     }
