+++ app/src/main/java/com/hldj/hmyg/saler/AdressListActivitt_bak.java
@@ -352,7 +352,7 @@ public class AdressListActivitt_bak extends NeedSwipeBackActivity implements IXL
             l_params = new LinearLayout.LayoutParams(LayoutParams.MATCH_PARENT,
                     LayoutParams.WRAP_CONTENT);
             WindowManager wm = ((Activity) context).getWindowManager();
-            l_params.height = (int) (wm.getDefaultDisplay().getWidth() * 9 / 20);
+            l_params.height = wm.getDefaultDisplay().getWidth() * 9 / 20;
         }
 
         @Override
