+++ app/src/main/java/com/hldj/hmyg/widget/ShareDialogFragment.java
@@ -92,7 +92,7 @@ public class ShareDialogFragment extends DialogFragment implements PlatformActio
         super.onStart();
         WindowManager.LayoutParams lp = getDialog().getWindow().getAttributes();
         lp.gravity = Gravity.BOTTOM;
-        lp.width = (int) (getDialog().getWindow().getWindowManager().getDefaultDisplay().getWidth());
+        lp.width = getDialog().getWindow().getWindowManager().getDefaultDisplay().getWidth();
         getDialog().getWindow().setGravity(Gravity.BOTTOM);
     }
